reorganize assets folder and remove staticimage
