+++ bikespace_frontend/src/assets/icons/icon.png
