+++ bikespace_frontend/src/components/submission/base-button/BaseButton.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
-import {StaticImage} from 'gatsby-plugin-image';
 
 import * as styles from './base-button.module.scss';
+
+import checkIcon from '@/assets/icons/check.svg';
 
 type BaseButtonProps = {
   type: string;
@@ -41,11 +42,7 @@ export class BaseButton extends React.Component<BaseButtonProps> {
         <label htmlFor={`${this.props.name}_${this.props.value}`}>
           <div className={styles.baseButton}>
             {this.props.children}
-            <StaticImage
-              className={styles.check}
-              src="../../../images/check.svg"
-              alt="checkmark"
-            />
+            <img className={styles.check} src={checkIcon} alt="checkmark" />
           </div>
         </label>
       </div>
