+++ bikespace_frontend/src/components/submission/submission-header/SubmissionHeader.tsx
@@ -1,17 +1,18 @@
 import React from 'react';
-import {StaticImage} from 'gatsby-plugin-image';
 
 import {FeedbackMailTo} from '../feedback-mail-to/FeedbackMailTo';
 
 import * as styles from './submission-header.module.scss';
 
+import bikespaceLogo from '@/assets/icons/header-logo.svg';
+
 export function SubmissionHeader() {
   return (
     <header className={styles.header}>
-      <StaticImage
+      <img
         className={styles.headerLogo}
-        src="../../../images/header-logo.svg"
-        alt="bike space logo"
+        src={bikespaceLogo}
+        alt="bikespace logo"
       />
       <FeedbackMailTo />
     </header>
