+++ bikespace_frontend/src/components/submission/submission-page/submission-page.module.scss
@@ -4,7 +4,7 @@
   main {
     position: relative;
     padding: 20px;
-    background-image: url('@/images/bikespace-background-desktop.jpeg');
+    background-image: url('@/assets/images/bikespace-background-desktop.jpeg');
     background-size: cover;
     background-position: center;
   }
